Correct syntax of helpers file; add current_user method
